requirement.txt
# Web & API Framework
fastapi
uvicorn

# Streamlit Frontend
streamlit

# Agents, Models & Tools
phidata
groq
google-generativeai
duckduckgo-search
yfinance

# Utilities
packaging
pypdf
python-dotenv
